improved performance removed optimize and added queued index through threaded executor.
== without performance tuning ==
first 100 items  ended in - 13.611 seconds.
== report ==
this time is not fixed, this is always increasing - i presume this is related with the total number of items already indexed.

== after removing optimization ==
first 100 items  ended in - 11.629 seconds.
== report ==
when i repeat this same test this time it is not changing or increasing like the previous scenario.

== after executor service for each repository instance ==
first 100 items  ended in - 5.6152 seconds.
== report ==
now this time is very repeating, and improved nearly 55%

now it should take 6,414 seconds to index 128281 items.
1.78 hours to index all

24,000
started on 3:08
ended on 3:26
taken : 18 mins